requirements.txt
python-telegram-bot>=21.6
cryptography>=43.0.0
httpx>=0.27.0
